fixup move global setup to tool_cfgable, make the main struct static
